+++ Components/Path.php
@@ -15,6 +15,7 @@ namespace League\Uri\Components;
 
 use Deprecated;
 use League\Uri\Contracts\PathInterface;
+use League\Uri\Contracts\UriException;
 use League\Uri\Contracts\UriInterface;
 use League\Uri\Encoder;
 use League\Uri\Uri;
@@ -57,6 +58,18 @@ final class Path extends Component implements PathInterface
     public static function new(Stringable|string $value = ''): self
     {
         return new self($value);
+    }
+
+    /**
+     * Create a new instance from a string.or a stringable structure or returns null on failure.
+     */
+    public static function tryNew(Stringable|string $uri = ''): ?self
+    {
+        try {
+            return self::new($uri);
+        } catch (UriException) {
+            return null;
+        }
     }
 
     /**