+++ Components/URLSearchParams.php
@@ -21,6 +21,7 @@ use Iterator;
 use IteratorAggregate;
 use League\Uri\Contracts\QueryInterface;
 use League\Uri\Contracts\UriComponentInterface;
+use League\Uri\Contracts\UriException;
 use League\Uri\Contracts\UriInterface;
 use League\Uri\Exceptions\SyntaxError;
 use League\Uri\KeyValuePair\Converter;
@@ -172,6 +173,18 @@ final class URLSearchParams implements Countable, IteratorAggregate, UriComponen
     public static function new(Stringable|string|null $query): self
     {
         return new self(Query::fromFormData(self::formatQuery($query)));
+    }
+
+    /**
+     * Create a new instance from a string.or a stringable structure or returns null on failure.
+     */
+    public static function tryNew(Stringable|string|null $uri = null): ?self
+    {
+        try {
+            return self::new($uri);
+        } catch (UriException) {
+            return null;
+        }
     }
 
     /**